# server/completions/completions.go
package completions

import (
	"code-completion/pkg/completions"
	"code-completion/pkg/model"
	"code-completion/pkg/stream_controller"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"
)

// Completions 补全接口路由处理
// @Summary 代码补全
// @Description 根据提供的代码上下文生成代码补全建议
// @Tags completions
// @Accept json
// @Produce json
// @Param request body completions.CompletionRequest true "补全请求"
// @Success 200 {object} completions.CompletionResponse
// @Failure 400 {object} map[string]interface{}
// @Failure 500 {object} map[string]interface{}
// @Router /api/completions [post]
// @Router /code-completion/api/v1/completions [post]
func Completions(c *gin.Context) {
	var req completions.CompletionInput
	if err := c.ShouldBindJSON(&req.CompletionRequest); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	req.Headers = c.Request.Header

	// 使用流控管理器处理请求
	if stream_controller.Controller != nil {
		rsp := stream_controller.Controller.ProcessCompletionRequest(c.Request.Context(), &req)
		respCompletion(c, &req.CompletionRequest, rsp)
		return
	}

	// 如果流控管理器未初始化，使用原有逻辑
	handler := completions.NewCompletionHandler(nil)
	perf := &completions.CompletionPerformance{
		ReceiveTime: time.Now().Local(),
	}
	rc := completions.NewCompletionContext(c.Request.Context(), perf)
	rsp := handler.HandleCompletion(rc, &req)
	respCompletion(c, &req.CompletionRequest, rsp)
}

func respCompletion(c *gin.Context, req *completions.CompletionRequest, rsp *completions.CompletionResponse) {
	if rsp.Status != model.StatusSuccess {
		zap.L().Warn("completion failed", zap.String("completionID", rsp.ID),
			zap.String("clientID", req.ClientID),
			zap.String("status", string(rsp.Status)),
			zap.Any("response", rsp))
	} else {
		zap.L().Info("completion succeeded", zap.String("completionID", rsp.ID),
			zap.String("clientID", req.ClientID),
			zap.Any("response", rsp))
	}
	statusCode := http.StatusOK
	// switch rsp.Status {
	// case model.StatusSuccess:
	// 	statusCode = http.StatusOK
	// case model.StatusCanceled:
	// 	statusCode = http.StatusRequestTimeout
	// case model.StatusTimeout:
	// 	statusCode = http.StatusGatewayTimeout
	// case model.StatusReqError:
	// case model.StatusRejected:
	// 	statusCode = http.StatusBadRequest
	// default:
	// 	statusCode = http.StatusInternalServerError
	// }
	c.JSON(statusCode, rsp)
}